Changes to lab files

+++ CPPLabs-and-Homework/Labs/Lab 09 Classes and Objects/Lab Problems/Problem 1 Coin Toss Simulator/main.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 int main()
 {
-    srand(static_cast<int>(time(0)));
+    srand((time(0)));
 
     Coin flipCoin;
